agents/requirements.txt
uagents==0.15.0
asyncio>=3.4.3
aiohttp>=3.8.0
python-dotenv>=1.0.0
web3>=6.0.0
requests>=2.28.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
cryptography>=3.4.8
